catchup

+++ internal/services/apimanagement/api_management_resource.go
@@ -33,9 +33,9 @@ import (
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/tf"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/features"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/apimanagement/schemaz"
 	apimValidate "github.com/hashicorp/terraform-provider-azurerm/internal/services/apimanagement/validate"
-	networkValidate "github.com/hashicorp/terraform-provider-azurerm/internal/services/network/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/validation"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/timeouts"
@@ -100,7 +100,7 @@ func resourceApiManagementService() *pluginsdk.Resource {
 }
 
 func resourceApiManagementSchema() map[string]*pluginsdk.Schema {
-	return map[string]*pluginsdk.Schema{
+	schema := map[string]*pluginsdk.Schema{
 		"name": schemaz.SchemaApiManagementName(),
 
 		"resource_group_name": commonschema.ResourceGroupName(),
@@ -212,7 +212,7 @@ func resourceApiManagementSchema() map[string]*pluginsdk.Schema {
 						ValidateFunc: validation.IntBetween(0, 12),
 					},
 
-					"zones": commonschema.ZonesMultipleOptionalForceNew(),
+					"zones": commonschema.ZonesMultipleOptional(),
 
 					"gateway_regional_url": {
 						Type:     pluginsdk.TypeString,
@@ -230,7 +230,7 @@ func resourceApiManagementSchema() map[string]*pluginsdk.Schema {
 					"public_ip_address_id": {
 						Type:         pluginsdk.TypeString,
 						Optional:     true,
-						ValidateFunc: networkValidate.PublicIpAddressID,
+						ValidateFunc: commonids.ValidatePublicIPAddressID,
 					},
 
 					"private_ip_addresses": {
@@ -463,32 +463,6 @@ func resourceApiManagementSchema() map[string]*pluginsdk.Schema {
 			},
 		},
 
-		// lintignore:XS003
-		"policy": {
-			Type:       pluginsdk.TypeList,
-			Optional:   true,
-			Computed:   true,
-			MaxItems:   1,
-			ConfigMode: pluginsdk.SchemaConfigModeAttr,
-			Elem: &pluginsdk.Resource{
-				Schema: map[string]*pluginsdk.Schema{
-					"xml_content": {
-						Type:             pluginsdk.TypeString,
-						Optional:         true,
-						Computed:         true,
-						ConflictsWith:    []string{"policy.0.xml_link"},
-						DiffSuppressFunc: XmlWithDotNetInterpolationsDiffSuppress,
-					},
-
-					"xml_link": {
-						Type:          pluginsdk.TypeString,
-						Optional:      true,
-						ConflictsWith: []string{"policy.0.xml_content"},
-					},
-				},
-			},
-		},
-
 		"sign_in": {
 			Type:     pluginsdk.TypeList,
 			Optional: true,
@@ -601,7 +575,7 @@ func resourceApiManagementSchema() map[string]*pluginsdk.Schema {
 		"public_ip_address_id": {
 			Type:         pluginsdk.TypeString,
 			Optional:     true,
-			ValidateFunc: networkValidate.PublicIpAddressID,
+			ValidateFunc: commonids.ValidatePublicIPAddressID,
 		},
 
 		"public_network_access_enabled": {
@@ -664,6 +638,36 @@ func resourceApiManagementSchema() map[string]*pluginsdk.Schema {
 
 		"tags": commonschema.Tags(),
 	}
+
+	if !features.FourPointOhBeta() {
+		schema["policy"] = &pluginsdk.Schema{
+			Type:       pluginsdk.TypeList,
+			Optional:   true,
+			Computed:   true,
+			MaxItems:   1,
+			ConfigMode: pluginsdk.SchemaConfigModeAttr,
+			Deprecated: "The `policy` block has been superseded by the resource `azurerm_api_management_policy` and will be removed in v4.0 of the AzureRM Provider",
+			Elem: &pluginsdk.Resource{
+				Schema: map[string]*pluginsdk.Schema{
+					"xml_content": {
+						Type:             pluginsdk.TypeString,
+						Optional:         true,
+						Computed:         true,
+						ConflictsWith:    []string{"policy.0.xml_link"},
+						DiffSuppressFunc: XmlWithDotNetInterpolationsDiffSuppress,
+					},
+
+					"xml_link": {
+						Type:          pluginsdk.TypeString,
+						Optional:      true,
+						ConflictsWith: []string{"policy.0.xml_content"},
+					},
+				},
+			},
+		}
+	}
+
+	return schema
 }
 
 func resourceApiManagementServiceCreate(d *pluginsdk.ResourceData, meta interface{}) error {
@@ -938,26 +942,28 @@ func resourceApiManagementServiceCreate(d *pluginsdk.ResourceData, meta interfac
 		}
 	}
 
-	policyClient := meta.(*clients.Client).ApiManagement.PolicyClient
-	policiesRaw := d.Get("policy").([]interface{})
-	policyContract, err := expandApiManagementPolicies(policiesRaw)
-	if err != nil {
-		return err
-	}
-
-	if _, ok := d.GetOk("policy"); ok {
-		policyServiceId := policy.NewServiceID(subscriptionId, id.ResourceGroupName, id.ServiceName)
-		// remove the existing policy
-		if delResp, err := policyClient.Delete(ctx, policyServiceId, policy.DeleteOperationOptions{}); err != nil {
-			if !response.WasNotFound(delResp.HttpResponse) {
-				return fmt.Errorf("removing Policies from %s: %+v", id, err)
-			}
+	if !features.FourPointOhBeta() {
+		policyClient := meta.(*clients.Client).ApiManagement.PolicyClient
+		policiesRaw := d.Get("policy").([]interface{})
+		policyContract, err := expandApiManagementPolicies(policiesRaw)
+		if err != nil {
+			return err
 		}
 
-		// then add the new one, if it exists
-		if policyContract != nil {
-			if _, err := policyClient.CreateOrUpdate(ctx, policyServiceId, *policyContract, policy.CreateOrUpdateOperationOptions{}); err != nil {
-				return fmt.Errorf(" setting Policies for %s: %+v", id, err)
+		if _, ok := d.GetOk("policy"); ok {
+			policyServiceId := policy.NewServiceID(subscriptionId, id.ResourceGroupName, id.ServiceName)
+			// remove the existing policy
+			if delResp, err := policyClient.Delete(ctx, policyServiceId, policy.DeleteOperationOptions{}); err != nil {
+				if !response.WasNotFound(delResp.HttpResponse) {
+					return fmt.Errorf("removing Policies from %s: %+v", id, err)
+				}
+			}
+
+			// then add the new one, if it exists
+			if policyContract != nil {
+				if _, err := policyClient.CreateOrUpdate(ctx, policyServiceId, *policyContract, policy.CreateOrUpdateOperationOptions{}); err != nil {
+					return fmt.Errorf(" setting Policies for %s: %+v", id, err)
+				}
 			}
 		}
 	}
@@ -1187,26 +1193,28 @@ func resourceApiManagementServiceUpdate(d *pluginsdk.ResourceData, meta interfac
 		}
 	}
 
-	if d.HasChange("policy") {
-		policyClient := meta.(*clients.Client).ApiManagement.PolicyClient
-		policiesRaw := d.Get("policy").([]interface{})
-		policyContract, err := expandApiManagementPolicies(policiesRaw)
-		if err != nil {
-			return err
-		}
-
-		policyServiceId := policy.NewServiceID(subscriptionId, id.ResourceGroupName, id.ServiceName)
-		// remove the existing policy
-		if delResp, err := policyClient.Delete(ctx, policyServiceId, policy.DeleteOperationOptions{}); err != nil {
-			if !response.WasNotFound(delResp.HttpResponse) {
-				return fmt.Errorf("removing Policies from %s: %+v", id, err)
+	if !features.FourPointOhBeta() {
+		if d.HasChange("policy") {
+			policyClient := meta.(*clients.Client).ApiManagement.PolicyClient
+			policiesRaw := d.Get("policy").([]interface{})
+			policyContract, err := expandApiManagementPolicies(policiesRaw)
+			if err != nil {
+				return err
 			}
-		}
 
-		// then add the new one, if it exists
-		if policyContract != nil {
-			if _, err := policyClient.CreateOrUpdate(ctx, policyServiceId, *policyContract, policy.CreateOrUpdateOperationOptions{}); err != nil {
-				return fmt.Errorf(" setting Policies for %s: %+v", id, err)
+			policyServiceId := policy.NewServiceID(subscriptionId, id.ResourceGroupName, id.ServiceName)
+			// remove the existing policy
+			if delResp, err := policyClient.Delete(ctx, policyServiceId, policy.DeleteOperationOptions{}); err != nil {
+				if !response.WasNotFound(delResp.HttpResponse) {
+					return fmt.Errorf("removing Policies from %s: %+v", id, err)
+				}
+			}
+
+			// then add the new one, if it exists
+			if policyContract != nil {
+				if _, err := policyClient.CreateOrUpdate(ctx, policyServiceId, *policyContract, policy.CreateOrUpdateOperationOptions{}); err != nil {
+					return fmt.Errorf(" setting Policies for %s: %+v", id, err)
+				}
 			}
 		}
 	}
@@ -1279,12 +1287,12 @@ func resourceApiManagementServiceRead(d *pluginsdk.ResourceData, meta interface{
 		d.Set("publisher_email", model.Properties.PublisherEmail)
 		d.Set("publisher_name", model.Properties.PublisherName)
 		d.Set("notification_sender_email", pointer.From(model.Properties.NotificationSenderEmail))
-		d.Set("gateway_url", pointer.From(model.Properties.GatewayUrl))
-		d.Set("gateway_regional_url", pointer.From(model.Properties.GatewayRegionalUrl))
-		d.Set("portal_url", pointer.From(model.Properties.PortalUrl))
-		d.Set("developer_portal_url", pointer.From(model.Properties.DeveloperPortalUrl))
-		d.Set("management_api_url", pointer.From(model.Properties.ManagementApiUrl))
-		d.Set("scm_url", pointer.From(model.Properties.ScmUrl))
+		d.Set("gateway_url", pointer.From(model.Properties.GatewayURL))
+		d.Set("gateway_regional_url", pointer.From(model.Properties.GatewayRegionalURL))
+		d.Set("portal_url", pointer.From(model.Properties.PortalURL))
+		d.Set("developer_portal_url", pointer.From(model.Properties.DeveloperPortalURL))
+		d.Set("management_api_url", pointer.From(model.Properties.ManagementApiURL))
+		d.Set("scm_url", pointer.From(model.Properties.ScmURL))
 		d.Set("public_ip_addresses", pointer.From(model.Properties.PublicIPAddresses))
 		d.Set("public_ip_address_id", pointer.From(model.Properties.PublicIPAddressId))
 		d.Set("public_network_access_enabled", pointer.From(model.Properties.PublicNetworkAccess) == apimanagementservice.PublicNetworkAccessEnabled)
@@ -1335,8 +1343,10 @@ func resourceApiManagementServiceRead(d *pluginsdk.ResourceData, meta interface{
 			return fmt.Errorf("setting `sku_name`: %+v", err)
 		}
 
-		if err := d.Set("policy", flattenApiManagementPolicies(d, policy.Model)); err != nil {
-			return fmt.Errorf("setting `policy`: %+v", err)
+		if !features.FourPointOhBeta() {
+			if err := d.Set("policy", flattenApiManagementPolicies(d, policy.Model)); err != nil {
+				return fmt.Errorf("setting `policy`: %+v", err)
+			}
 		}
 
 		d.Set("zones", zones.FlattenUntyped(model.Zones))
@@ -1729,7 +1739,7 @@ func flattenApiManagementAdditionalLocations(input *[]apimanagementservice.Addit
 
 		results = append(results, map[string]interface{}{
 			"capacity":                      int32(prop.Sku.Capacity),
-			"gateway_regional_url":          pointer.From(prop.GatewayRegionalUrl),
+			"gateway_regional_url":          pointer.From(prop.GatewayRegionalURL),
 			"location":                      location.NormalizeNilable(pointer.To(prop.Location)),
 			"private_ip_addresses":          pointer.From(prop.PrivateIPAddresses),
 			"public_ip_address_id":          pointer.From(prop.PublicIPAddressId),
@@ -2174,7 +2184,7 @@ func flattenApiManagementPolicies(d *pluginsdk.ResourceData, input *policy.Polic
 	// as such we need to retrieve this value from the state if it's present
 	if existing, ok := d.GetOk("policy"); ok {
 		existingVs := existing.([]interface{})
-		if len(existingVs) > 0 {
+		if len(existingVs) > 0 && existingVs[0] != nil {
 			existingV := existingVs[0].(map[string]interface{})
 			output["xml_link"] = existingV["xml_link"].(string)
 		}
